requirements.txt
aiogram==3.15.0
python-dotenv==1.0.1
aiofiles==23.2.1
aiohttp==3.10.11
asyncio-throttle==1.0.2 